refactor(options): move option callbacks from option.c to option_cb.c
Problem: `option.c` is too big

Solution: Move option callbacks from `option.c` to `option_cb.c`.
A possible follow-up PR would be to move option callbacks from `optionstr.c` to `option_cb.c` as well.
