+++ src/nvim/option.c
@@ -122,25 +122,9 @@ static const char e_not_allowed_in_modeline_when_modelineexpr_is_off[]
   = N_("E992: Not allowed in a modeline when 'modelineexpr' is off");
 static const char e_number_required_after_equal[]
   = N_("E521: Number required after =");
-static const char e_preview_window_already_exists[]
-  = N_("E590: A preview window already exists");
 
 static char *p_term = NULL;
 static char *p_ttytype = NULL;
-
-// Saved values for when 'bin' is set.
-static int p_et_nobin;
-static int p_ml_nobin;
-static OptInt p_tw_nobin;
-static OptInt p_wm_nobin;
-
-// Saved values for when 'paste' is set.
-static int p_ai_nopaste;
-static int p_et_nopaste;
-static OptInt p_sts_nopaste;
-static OptInt p_tw_nopaste;
-static OptInt p_wm_nopaste;
-static char *p_vsts_nopaste;
 
 #define OPTION_COUNT ARRAY_SIZE(options)
 
@@ -163,15 +147,6 @@ typedef enum {
 # include "options.generated.h"
 # include "options_map.generated.h"
 #endif
-
-static int p_bin_dep_opts[] = {
-  kOptTextwidth, kOptWrapmargin, kOptModeline, kOptExpandtab, kOptInvalid
-};
-
-static int p_paste_dep_opts[] = {
-  kOptAutoindent, kOptExpandtab, kOptRuler, kOptShowmatch, kOptSmarttab, kOptSofttabstop,
-  kOptTextwidth, kOptWrapmargin, kOptRevins, kOptVarsofttabstop, kOptInvalid
-};
 
 void set_init_tablocal(void)
 {
@@ -1479,72 +1454,6 @@ int string_to_key(char *arg)
   return (uint8_t)(*arg);
 }
 
-// When changing 'title', 'titlestring', 'icon' or 'iconstring', call
-// maketitle() to create and display it.
-// When switching the title or icon off, call ui_set_{icon,title}(NULL) to get
-// the old value back.
-void did_set_title(void)
-{
-  if (starting != NO_SCREEN) {
-    maketitle();
-  }
-}
-
-/// set_options_bin -  called when 'bin' changes value.
-///
-/// @param opt_flags  OPT_LOCAL and/or OPT_GLOBAL
-void set_options_bin(int oldval, int newval, int opt_flags)
-{
-  // The option values that are changed when 'bin' changes are
-  // copied when 'bin is set and restored when 'bin' is reset.
-  if (newval) {
-    if (!oldval) {              // switched on
-      if (!(opt_flags & OPT_GLOBAL)) {
-        curbuf->b_p_tw_nobin = curbuf->b_p_tw;
-        curbuf->b_p_wm_nobin = curbuf->b_p_wm;
-        curbuf->b_p_ml_nobin = curbuf->b_p_ml;
-        curbuf->b_p_et_nobin = curbuf->b_p_et;
-      }
-      if (!(opt_flags & OPT_LOCAL)) {
-        p_tw_nobin = p_tw;
-        p_wm_nobin = p_wm;
-        p_ml_nobin = p_ml;
-        p_et_nobin = p_et;
-      }
-    }
-
-    if (!(opt_flags & OPT_GLOBAL)) {
-      curbuf->b_p_tw = 0;       // no automatic line wrap
-      curbuf->b_p_wm = 0;       // no automatic line wrap
-      curbuf->b_p_ml = 0;       // no modelines
-      curbuf->b_p_et = 0;       // no expandtab
-    }
-    if (!(opt_flags & OPT_LOCAL)) {
-      p_tw = 0;
-      p_wm = 0;
-      p_ml = false;
-      p_et = false;
-      p_bin = true;             // needed when called for the "-b" argument
-    }
-  } else if (oldval) {        // switched off
-    if (!(opt_flags & OPT_GLOBAL)) {
-      curbuf->b_p_tw = curbuf->b_p_tw_nobin;
-      curbuf->b_p_wm = curbuf->b_p_wm_nobin;
-      curbuf->b_p_ml = curbuf->b_p_ml_nobin;
-      curbuf->b_p_et = curbuf->b_p_et_nobin;
-    }
-    if (!(opt_flags & OPT_LOCAL)) {
-      p_tw = p_tw_nobin;
-      p_wm = p_wm_nobin;
-      p_ml = p_ml_nobin;
-      p_et = p_et_nobin;
-    }
-  }
-
-  // Remember where the dependent option were reset
-  didset_options_sctx(opt_flags, p_bin_dep_opts);
-}
-
 /// Find the parameter represented by the given character (eg ', :, ", or /),
 /// and return its associated value in the 'shada' string.
 /// Only works for number parameters, not for 'r' or 'n'.
@@ -1887,831 +1796,6 @@ static void apply_optionset_autocmd(OptIndex opt_idx, int opt_flags, OptVal oldv
   }
   apply_autocmds(EVENT_OPTIONSET, options[opt_idx].fullname, NULL, false, NULL);
   reset_v_option_vars();
-}
-
-/// Process the updated 'arabic' option value.
-static const char *did_set_arabic(optset_T *args)
-{
-  win_T *win = (win_T *)args->os_win;
-  const char *errmsg = NULL;
-
-  if (win->w_p_arab) {
-    // 'arabic' is set, handle various sub-settings.
-    if (!p_tbidi) {
-      // set rightleft mode
-      if (!win->w_p_rl) {
-        win->w_p_rl = true;
-        changed_window_setting(curwin);
-      }
-
-      // Enable Arabic shaping (major part of what Arabic requires)
-      if (!p_arshape) {
-        p_arshape = true;
-        redraw_all_later(UPD_NOT_VALID);
-      }
-    }
-
-    // Arabic requires a utf-8 encoding, inform the user if it's not
-    // set.
-    if (strcmp(p_enc, "utf-8") != 0) {
-      static char *w_arabic = N_("W17: Arabic requires UTF-8, do ':set encoding=utf-8'");
-
-      msg_source(HL_ATTR(HLF_W));
-      msg(_(w_arabic), HL_ATTR(HLF_W));
-      set_vim_var_string(VV_WARNINGMSG, _(w_arabic), -1);
-    }
-
-    // set 'delcombine'
-    p_deco = true;
-
-    // Force-set the necessary keymap for arabic.
-    errmsg = set_option_value(kOptKeymap, STATIC_CSTR_AS_OPTVAL("arabic"), OPT_LOCAL);
-  } else {
-    // 'arabic' is reset, handle various sub-settings.
-    if (!p_tbidi) {
-      // reset rightleft mode
-      if (win->w_p_rl) {
-        win->w_p_rl = false;
-        changed_window_setting(curwin);
-      }
-
-      // 'arabicshape' isn't reset, it is a global option and
-      // another window may still need it "on".
-    }
-
-    // 'delcombine' isn't reset, it is a global option and another
-    // window may still want it "on".
-
-    // Revert to the default keymap
-    curbuf->b_p_iminsert = B_IMODE_NONE;
-    curbuf->b_p_imsearch = B_IMODE_USE_INSERT;
-  }
-
-  return errmsg;
-}
-
-/// Process the updated 'autochdir' option value.
-static const char *did_set_autochdir(optset_T *args FUNC_ATTR_UNUSED)
-{
-  // Change directories when the 'acd' option is set now.
-  do_autochdir();
-  return NULL;
-}
-
-/// Process the updated 'binary' option value.
-static const char *did_set_binary(optset_T *args)
-{
-  buf_T *buf = (buf_T *)args->os_buf;
-
-  // when 'bin' is set also set some other options
-  set_options_bin((int)args->os_oldval.boolean, buf->b_p_bin, args->os_flags);
-  redraw_titles();
-
-  return NULL;
-}
-
-/// Process the updated 'buflisted' option value.
-static const char *did_set_buflisted(optset_T *args)
-{
-  buf_T *buf = (buf_T *)args->os_buf;
-
-  // when 'buflisted' changes, trigger autocommands
-  if (args->os_oldval.boolean != buf->b_p_bl) {
-    apply_autocmds(buf->b_p_bl ? EVENT_BUFADD : EVENT_BUFDELETE,
-                   NULL, NULL, true, buf);
-  }
-  return NULL;
-}
-
-/// Process the new 'cmdheight' option value.
-static const char *did_set_cmdheight(optset_T *args)
-{
-  OptInt old_value = args->os_oldval.number;
-
-  if (p_ch > Rows - min_rows() + 1) {
-    p_ch = Rows - min_rows() + 1;
-  }
-
-  // if p_ch changed value, change the command line height
-  // Only compute the new window layout when startup has been
-  // completed. Otherwise the frame sizes may be wrong.
-  if ((p_ch != old_value
-       || tabline_height() + global_stl_height() + topframe->fr_height != Rows - p_ch)
-      && full_screen) {
-    command_height();
-  }
-
-  return NULL;
-}
-
-/// Process the updated 'diff' option value.
-static const char *did_set_diff(optset_T *args)
-{
-  win_T *win = (win_T *)args->os_win;
-  // May add or remove the buffer from the list of diff buffers.
-  diff_buf_adjust(win);
-  if (foldmethodIsDiff(win)) {
-    foldUpdateAll(win);
-  }
-  return NULL;
-}
-
-/// Process the updated 'endoffile' or 'endofline' or 'fixendofline' or 'bomb'
-/// option value.
-static const char *did_set_eof_eol_fixeol_bomb(optset_T *args FUNC_ATTR_UNUSED)
-{
-  // redraw the window title and tab page text
-  redraw_titles();
-  return NULL;
-}
-
-/// Process the updated 'equalalways' option value.
-static const char *did_set_equalalways(optset_T *args)
-{
-  win_T *win = (win_T *)args->os_win;
-  if (p_ea && !args->os_oldval.boolean) {
-    win_equal(win, false, 0);
-  }
-
-  return NULL;
-}
-
-/// Process the new 'foldlevel' option value.
-static const char *did_set_foldlevel(optset_T *args FUNC_ATTR_UNUSED)
-{
-  newFoldLevel();
-  return NULL;
-}
-
-/// Process the new 'foldminlines' option value.
-static const char *did_set_foldminlines(optset_T *args)
-{
-  win_T *win = (win_T *)args->os_win;
-  foldUpdateAll(win);
-  return NULL;
-}
-
-/// Process the new 'foldnestmax' option value.
-static const char *did_set_foldnestmax(optset_T *args)
-{
-  win_T *win = (win_T *)args->os_win;
-  if (foldmethodIsSyntax(win) || foldmethodIsIndent(win)) {
-    foldUpdateAll(win);
-  }
-  return NULL;
-}
-
-/// Process the new 'helpheight' option value.
-static const char *did_set_helpheight(optset_T *args)
-{
-  // Change window height NOW
-  if (!ONE_WINDOW) {
-    buf_T *buf = (buf_T *)args->os_buf;
-    win_T *win = (win_T *)args->os_win;
-    if (buf->b_help && win->w_height < p_hh) {
-      win_setheight((int)p_hh);
-    }
-  }
-
-  return NULL;
-}
-
-/// Process the updated 'hlsearch' option value.
-static const char *did_set_hlsearch(optset_T *args FUNC_ATTR_UNUSED)
-{
-  // when 'hlsearch' is set or reset: reset no_hlsearch
-  set_no_hlsearch(false);
-  return NULL;
-}
-
-/// Process the updated 'ignorecase' option value.
-static const char *did_set_ignorecase(optset_T *args FUNC_ATTR_UNUSED)
-{
-  // when 'ignorecase' is set or reset and 'hlsearch' is set, redraw
-  if (p_hls) {
-    redraw_all_later(UPD_SOME_VALID);
-  }
-  return NULL;
-}
-
-/// Process the new 'iminset' option value.
-static const char *did_set_iminsert(optset_T *args FUNC_ATTR_UNUSED)
-{
-  showmode();
-  // Show/unshow value of 'keymap' in status lines.
-  status_redraw_curbuf();
-
-  return NULL;
-}
-
-/// Process the updated 'langnoremap' option value.
-static const char *did_set_langnoremap(optset_T *args FUNC_ATTR_UNUSED)
-{
-  // 'langnoremap' -> !'langremap'
-  p_lrm = !p_lnr;
-  return NULL;
-}
-
-/// Process the updated 'langremap' option value.
-static const char *did_set_langremap(optset_T *args FUNC_ATTR_UNUSED)
-{
-  // 'langremap' -> !'langnoremap'
-  p_lnr = !p_lrm;
-  return NULL;
-}
-
-/// Process the new 'laststatus' option value.
-static const char *did_set_laststatus(optset_T *args)
-{
-  OptInt old_value = args->os_oldval.number;
-  OptInt value = args->os_newval.number;
-
-  // When switching to global statusline, decrease topframe height
-  // Also clear the cmdline to remove the ruler if there is one
-  if (value == 3 && old_value != 3) {
-    frame_new_height(topframe, topframe->fr_height - STATUS_HEIGHT, false, false);
-    win_comp_pos();
-    clear_cmdline = true;
-  }
-  // When switching from global statusline, increase height of topframe by STATUS_HEIGHT
-  // in order to to re-add the space that was previously taken by the global statusline
-  if (old_value == 3 && value != 3) {
-    frame_new_height(topframe, topframe->fr_height + STATUS_HEIGHT, false, false);
-    win_comp_pos();
-  }
-
-  last_status(false);  // (re)set last window status line.
-  return NULL;
-}
-
-/// Process the updated 'lines' or 'columns' option value.
-static const char *did_set_lines_or_columns(optset_T *args)
-{
-  // If the screen (shell) height has been changed, assume it is the
-  // physical screenheight.
-  if (p_lines != Rows || p_columns != Columns) {
-    // Changing the screen size is not allowed while updating the screen.
-    if (updating_screen) {
-      OptVal oldval = (OptVal){ .type = kOptValTypeNumber, .data = args->os_oldval };
-      set_option_varp(args->os_idx, args->os_varp, oldval, false);
-    } else if (full_screen) {
-      screen_resize((int)p_columns, (int)p_lines);
-    } else {
-      // TODO(bfredl): is this branch ever needed?
-      // Postpone the resizing; check the size and cmdline position for
-      // messages.
-      Rows = (int)p_lines;
-      Columns = (int)p_columns;
-      check_screensize();
-      int new_row = (int)(Rows - MAX(p_ch, 1));
-      if (cmdline_row > new_row && Rows > p_ch) {
-        assert(p_ch >= 0 && new_row <= INT_MAX);
-        cmdline_row = new_row;
-      }
-    }
-    if (p_window >= Rows || !option_was_set(kOptWindow)) {
-      p_window = Rows - 1;
-    }
-  }
-
-  // Adjust 'scrolljump' if needed.
-  if (p_sj >= Rows && full_screen) {
-    p_sj = Rows / 2;
-  }
-
-  return NULL;
-}
-
-/// Process the updated 'lisp' option value.
-static const char *did_set_lisp(optset_T *args)
-{
-  buf_T *buf = (buf_T *)args->os_buf;
-  // When 'lisp' option changes include/exclude '-' in keyword characters.
-  buf_init_chartab(buf, false);          // ignore errors
-  return NULL;
-}
-
-/// Process the updated 'modifiable' option value.
-static const char *did_set_modifiable(optset_T *args FUNC_ATTR_UNUSED)
-{
-  // when 'modifiable' is changed, redraw the window title
-  redraw_titles();
-
-  return NULL;
-}
-
-/// Process the updated 'modified' option value.
-static const char *did_set_modified(optset_T *args)
-{
-  buf_T *buf = (buf_T *)args->os_buf;
-  if (!args->os_newval.boolean) {
-    save_file_ff(buf);  // Buffer is unchanged
-  }
-  redraw_titles();
-  buf->b_modified_was_set = (int)args->os_newval.boolean;
-  return NULL;
-}
-
-/// Process the updated 'number' or 'relativenumber' option value.
-static const char *did_set_number_relativenumber(optset_T *args)
-{
-  win_T *win = (win_T *)args->os_win;
-  if (*win->w_p_stc != NUL) {
-    // When 'relativenumber'/'number' is changed and 'statuscolumn' is set, reset width.
-    win->w_nrwidth_line_count = 0;
-  }
-  check_signcolumn(NULL, win);
-  return NULL;
-}
-
-/// Process the new 'numberwidth' option value.
-static const char *did_set_numberwidth(optset_T *args)
-{
-  win_T *win = (win_T *)args->os_win;
-  win->w_nrwidth_line_count = 0;  // trigger a redraw
-
-  return NULL;
-}
-
-/// Process the updated 'paste' option value.
-static const char *did_set_paste(optset_T *args FUNC_ATTR_UNUSED)
-{
-  static int old_p_paste = false;
-  static int save_sm = 0;
-  static int save_sta = 0;
-  static int save_ru = 0;
-  static int save_ri = 0;
-
-  if (p_paste) {
-    // Paste switched from off to on.
-    // Save the current values, so they can be restored later.
-    if (!old_p_paste) {
-      // save options for each buffer
-      FOR_ALL_BUFFERS(buf) {
-        buf->b_p_tw_nopaste = buf->b_p_tw;
-        buf->b_p_wm_nopaste = buf->b_p_wm;
-        buf->b_p_sts_nopaste = buf->b_p_sts;
-        buf->b_p_ai_nopaste = buf->b_p_ai;
-        buf->b_p_et_nopaste = buf->b_p_et;
-        if (buf->b_p_vsts_nopaste) {
-          xfree(buf->b_p_vsts_nopaste);
-        }
-        buf->b_p_vsts_nopaste = buf->b_p_vsts && buf->b_p_vsts != empty_string_option
-                                ? xstrdup(buf->b_p_vsts)
-                                : NULL;
-      }
-
-      // save global options
-      save_sm = p_sm;
-      save_sta = p_sta;
-      save_ru = p_ru;
-      save_ri = p_ri;
-      // save global values for local buffer options
-      p_ai_nopaste = p_ai;
-      p_et_nopaste = p_et;
-      p_sts_nopaste = p_sts;
-      p_tw_nopaste = p_tw;
-      p_wm_nopaste = p_wm;
-      if (p_vsts_nopaste) {
-        xfree(p_vsts_nopaste);
-      }
-      p_vsts_nopaste = p_vsts && p_vsts != empty_string_option ? xstrdup(p_vsts) : NULL;
-    }
-
-    // Always set the option values, also when 'paste' is set when it is
-    // already on.
-    // set options for each buffer
-    FOR_ALL_BUFFERS(buf) {
-      buf->b_p_tw = 0;              // textwidth is 0
-      buf->b_p_wm = 0;              // wrapmargin is 0
-      buf->b_p_sts = 0;             // softtabstop is 0
-      buf->b_p_ai = 0;              // no auto-indent
-      buf->b_p_et = 0;              // no expandtab
-      if (buf->b_p_vsts) {
-        free_string_option(buf->b_p_vsts);
-      }
-      buf->b_p_vsts = empty_string_option;
-      XFREE_CLEAR(buf->b_p_vsts_array);
-    }
-
-    // set global options
-    p_sm = 0;                       // no showmatch
-    p_sta = 0;                      // no smarttab
-    if (p_ru) {
-      status_redraw_all();          // redraw to remove the ruler
-    }
-    p_ru = 0;                       // no ruler
-    p_ri = 0;                       // no reverse insert
-    // set global values for local buffer options
-    p_tw = 0;
-    p_wm = 0;
-    p_sts = 0;
-    p_ai = 0;
-    p_et = 0;
-    if (p_vsts) {
-      free_string_option(p_vsts);
-    }
-    p_vsts = empty_string_option;
-  } else if (old_p_paste) {
-    // Paste switched from on to off: Restore saved values.
-
-    // restore options for each buffer
-    FOR_ALL_BUFFERS(buf) {
-      buf->b_p_tw = buf->b_p_tw_nopaste;
-      buf->b_p_wm = buf->b_p_wm_nopaste;
-      buf->b_p_sts = buf->b_p_sts_nopaste;
-      buf->b_p_ai = buf->b_p_ai_nopaste;
-      buf->b_p_et = buf->b_p_et_nopaste;
-      if (buf->b_p_vsts) {
-        free_string_option(buf->b_p_vsts);
-      }
-      buf->b_p_vsts = buf->b_p_vsts_nopaste ? xstrdup(buf->b_p_vsts_nopaste) : empty_string_option;
-      xfree(buf->b_p_vsts_array);
-      if (buf->b_p_vsts && buf->b_p_vsts != empty_string_option) {
-        tabstop_set(buf->b_p_vsts, &buf->b_p_vsts_array);
-      } else {
-        buf->b_p_vsts_array = NULL;
-      }
-    }
-
-    // restore global options
-    p_sm = save_sm;
-    p_sta = save_sta;
-    if (p_ru != save_ru) {
-      status_redraw_all();          // redraw to draw the ruler
-    }
-    p_ru = save_ru;
-    p_ri = save_ri;
-    // set global values for local buffer options
-    p_ai = p_ai_nopaste;
-    p_et = p_et_nopaste;
-    p_sts = p_sts_nopaste;
-    p_tw = p_tw_nopaste;
-    p_wm = p_wm_nopaste;
-    if (p_vsts) {
-      free_string_option(p_vsts);
-    }
-    p_vsts = p_vsts_nopaste ? xstrdup(p_vsts_nopaste) : empty_string_option;
-  }
-
-  old_p_paste = p_paste;
-
-  // Remember where the dependent options were reset
-  didset_options_sctx((OPT_LOCAL | OPT_GLOBAL), p_paste_dep_opts);
-
-  return NULL;
-}
-
-/// Process the updated 'previewwindow' option value.
-static const char *did_set_previewwindow(optset_T *args)
-{
-  win_T *win = (win_T *)args->os_win;
-
-  if (!win->w_p_pvw) {
-    return NULL;
-  }
-
-  // There can be only one window with 'previewwindow' set.
-  FOR_ALL_WINDOWS_IN_TAB(wp, curtab) {
-    if (wp->w_p_pvw && wp != win) {
-      win->w_p_pvw = false;
-      return e_preview_window_already_exists;
-    }
-  }
-
-  return NULL;
-}
-
-/// Process the new 'pumblend' option value.
-static const char *did_set_pumblend(optset_T *args FUNC_ATTR_UNUSED)
-{
-  hl_invalidate_blends();
-  pum_grid.blending = (p_pb > 0);
-  if (pum_drawn()) {
-    pum_redraw();
-  }
-
-  return NULL;
-}
-
-/// Process the updated 'readonly' option value.
-static const char *did_set_readonly(optset_T *args)
-{
-  // when 'readonly' is reset globally, also reset readonlymode
-  if (!curbuf->b_p_ro && (args->os_flags & OPT_LOCAL) == 0) {
-    readonlymode = false;
-  }
-
-  // when 'readonly' is set may give W10 again
-  if (curbuf->b_p_ro) {
-    curbuf->b_did_warn = false;
-  }
-
-  redraw_titles();
-
-  return NULL;
-}
-
-/// Process the new 'scrollback' option value.
-static const char *did_set_scrollback(optset_T *args)
-{
-  buf_T *buf = (buf_T *)args->os_buf;
-  OptInt old_value = args->os_oldval.number;
-  OptInt value = args->os_newval.number;
-
-  if (buf->terminal && value < old_value) {
-    // Force the scrollback to take immediate effect only when decreasing it.
-    on_scrollback_option_changed(buf->terminal);
-  }
-  return NULL;
-}
-
-/// Process the updated 'scrollbind' option value.
-static const char *did_set_scrollbind(optset_T *args)
-{
-  win_T *win = (win_T *)args->os_win;
-
-  // when 'scrollbind' is set: snapshot the current position to avoid a jump
-  // at the end of normal_cmd()
-  if (!win->w_p_scb) {
-    return NULL;
-  }
-  do_check_scrollbind(false);
-  win->w_scbind_pos = get_vtopline(win);
-  return NULL;
-}
-
-#ifdef BACKSLASH_IN_FILENAME
-/// Process the updated 'shellslash' option value.
-static const char *did_set_shellslash(optset_T *args FUNC_ATTR_UNUSED)
-{
-  if (p_ssl) {
-    psepc = '/';
-    psepcN = '\\';
-    pseps[0] = '/';
-  } else {
-    psepc = '\\';
-    psepcN = '/';
-    pseps[0] = '\\';
-  }
-
-  // need to adjust the file name arguments and buffer names.
-  buflist_slash_adjust();
-  alist_slash_adjust();
-  scriptnames_slash_adjust();
-  return NULL;
-}
-#endif
-
-/// Process the new 'shiftwidth' or the 'tabstop' option value.
-static const char *did_set_shiftwidth_tabstop(optset_T *args)
-{
-  buf_T *buf = (buf_T *)args->os_buf;
-  win_T *win = (win_T *)args->os_win;
-  OptInt *pp = (OptInt *)args->os_varp;
-
-  if (foldmethodIsIndent(win)) {
-    foldUpdateAll(win);
-  }
-  // When 'shiftwidth' changes, or it's zero and 'tabstop' changes:
-  // parse 'cinoptions'.
-  if (pp == &buf->b_p_sw || buf->b_p_sw == 0) {
-    parse_cino(buf);
-  }
-
-  return NULL;
-}
-
-/// Process the new 'showtabline' option value.
-static const char *did_set_showtabline(optset_T *args FUNC_ATTR_UNUSED)
-{
-  // (re)set tab page line
-  win_new_screen_rows();  // recompute window positions and heights
-  return NULL;
-}
-
-/// Process the updated 'smoothscroll' option value.
-static const char *did_set_smoothscroll(optset_T *args FUNC_ATTR_UNUSED)
-{
-  win_T *win = (win_T *)args->os_win;
-  if (!win->w_p_sms) {
-    win->w_skipcol = 0;
-  }
-
-  return NULL;
-}
-
-/// Process the updated 'spell' option value.
-static const char *did_set_spell(optset_T *args)
-{
-  win_T *win = (win_T *)args->os_win;
-  if (win->w_p_spell) {
-    return parse_spelllang(win);
-  }
-
-  return NULL;
-}
-
-/// Process the updated 'swapfile' option value.
-static const char *did_set_swapfile(optset_T *args)
-{
-  buf_T *buf = (buf_T *)args->os_buf;
-  // when 'swf' is set, create swapfile, when reset remove swapfile
-  if (buf->b_p_swf && p_uc) {
-    ml_open_file(buf);                     // create the swap file
-  } else {
-    // no need to reset curbuf->b_may_swap, ml_open_file() will check
-    // buf->b_p_swf
-    mf_close_file(buf, true);              // remove the swap file
-  }
-  return NULL;
-}
-
-/// Process the new 'textwidth' option value.
-static const char *did_set_textwidth(optset_T *args FUNC_ATTR_UNUSED)
-{
-  FOR_ALL_TAB_WINDOWS(tp, wp) {
-    check_colorcolumn(NULL, wp);
-  }
-
-  return NULL;
-}
-
-/// Process the updated 'title' or the 'icon' option value.
-static const char *did_set_title_icon(optset_T *args FUNC_ATTR_UNUSED)
-{
-  // when 'title' changed, may need to change the title; same for 'icon'
-  did_set_title();
-  return NULL;
-}
-
-/// Process the new 'titlelen' option value.
-static const char *did_set_titlelen(optset_T *args)
-{
-  OptInt old_value = args->os_oldval.number;
-
-  // if 'titlelen' has changed, redraw the title
-  if (starting != NO_SCREEN && old_value != p_titlelen) {
-    need_maketitle = true;
-  }
-
-  return NULL;
-}
-
-/// Process the updated 'undofile' option value.
-static const char *did_set_undofile(optset_T *args)
-{
-  // Only take action when the option was set.
-  if (!curbuf->b_p_udf && !p_udf) {
-    return NULL;
-  }
-
-  // When reset we do not delete the undo file, the option may be set again
-  // without making any changes in between.
-  uint8_t hash[UNDO_HASH_SIZE];
-
-  FOR_ALL_BUFFERS(bp) {
-    // When 'undofile' is set globally: for every buffer, otherwise
-    // only for the current buffer: Try to read in the undofile,
-    // if one exists, the buffer wasn't changed and the buffer was
-    // loaded
-    if ((curbuf == bp
-         || (args->os_flags & OPT_GLOBAL) || args->os_flags == 0)
-        && !bufIsChanged(bp) && bp->b_ml.ml_mfp != NULL) {
-      u_compute_hash(bp, hash);
-      u_read_undo(NULL, hash, bp->b_fname);
-    }
-  }
-
-  return NULL;
-}
-
-/// Process the new global 'undolevels' option value.
-const char *did_set_global_undolevels(OptInt value, OptInt old_value)
-{
-  // sync undo before 'undolevels' changes
-  // use the old value, otherwise u_sync() may not work properly
-  p_ul = old_value;
-  u_sync(true);
-  p_ul = value;
-  return NULL;
-}
-
-/// Process the new buffer local 'undolevels' option value.
-const char *did_set_buflocal_undolevels(buf_T *buf, OptInt value, OptInt old_value)
-{
-  // use the old value, otherwise u_sync() may not work properly
-  buf->b_p_ul = old_value;
-  u_sync(true);
-  buf->b_p_ul = value;
-  return NULL;
-}
-
-/// Process the new 'undolevels' option value.
-static const char *did_set_undolevels(optset_T *args)
-{
-  buf_T *buf = (buf_T *)args->os_buf;
-  OptInt *pp = (OptInt *)args->os_varp;
-
-  if (pp == &p_ul) {                  // global 'undolevels'
-    did_set_global_undolevels(args->os_newval.number, args->os_oldval.number);
-  } else if (pp == &curbuf->b_p_ul) {      // buffer local 'undolevels'
-    did_set_buflocal_undolevels(buf, args->os_newval.number, args->os_oldval.number);
-  }
-
-  return NULL;
-}
-
-/// Process the new 'updatecount' option value.
-static const char *did_set_updatecount(optset_T *args)
-{
-  OptInt old_value = args->os_oldval.number;
-
-  // when 'updatecount' changes from zero to non-zero, open swap files
-  if (p_uc && !old_value) {
-    ml_open_files();
-  }
-
-  return NULL;
-}
-
-/// Process the new 'wildchar' / 'wildcharm' option value.
-static const char *did_set_wildchar(optset_T *args)
-{
-  OptInt c = *(OptInt *)args->os_varp;
-
-  // Don't allow key values that wouldn't work as wildchar.
-  if (c == Ctrl_C || c == '\n' || c == '\r' || c == K_KENTER) {
-    return e_invarg;
-  }
-
-  return NULL;
-}
-
-/// Process the new 'winblend' option value.
-static const char *did_set_winblend(optset_T *args)
-{
-  win_T *win = (win_T *)args->os_win;
-  OptInt old_value = args->os_oldval.number;
-  OptInt value = args->os_newval.number;
-
-  if (value != old_value) {
-    win->w_p_winbl = MAX(MIN(win->w_p_winbl, 100), 0);
-    win->w_hl_needs_update = true;
-    check_blending(win);
-  }
-
-  return NULL;
-}
-
-/// Process the new 'window' option value.
-static const char *did_set_window(optset_T *args FUNC_ATTR_UNUSED)
-{
-  if (p_window < 1) {
-    p_window = Rows - 1;
-  } else if (p_window >= Rows) {
-    p_window = Rows - 1;
-  }
-  return NULL;
-}
-
-/// Process the new 'winheight' value.
-static const char *did_set_winheight(optset_T *args)
-{
-  // Change window height NOW
-  if (!ONE_WINDOW) {
-    win_T *win = (win_T *)args->os_win;
-    if (win->w_height < p_wh) {
-      win_setheight((int)p_wh);
-    }
-  }
-
-  return NULL;
-}
-
-/// Process the new 'winwidth' option value.
-static const char *did_set_winwidth(optset_T *args)
-{
-  win_T *win = (win_T *)args->os_win;
-
-  if (!ONE_WINDOW && win->w_width < p_wiw) {
-    win_setwidth((int)p_wiw);
-  }
-  return NULL;
-}
-
-/// Process the updated 'wrap' option value.
-static const char *did_set_wrap(optset_T *args)
-{
-  win_T *win = (win_T *)args->os_win;
-  // Set w_leftcol or w_skipcol to zero.
-  if (win->w_p_wrap) {
-    win->w_leftcol = 0;
-  } else {
-    win->w_skipcol = 0;
-  }
-
-  return NULL;
 }
 
 // When 'syntax' is set, load the syntax of that name
@@ -3193,7 +2277,7 @@ OptVal optval_from_varp(OptIndex opt_idx, void *varp)
 /// @param[out]  varp         Pointer to option variable.
 /// @param[in]   value        New option value.
 /// @param       free_oldval  Free old value.
-static void set_option_varp(OptIndex opt_idx, void *varp, OptVal value, bool free_oldval)
+void set_option_varp(OptIndex opt_idx, void *varp, OptVal value, bool free_oldval)
   FUNC_ATTR_NONNULL_ARG(2)
 {
   assert(option_has_type(opt_idx, value.type));
@@ -6171,17 +5255,6 @@ int option_set_callback_func(char *optval, Callback *optcb)
   *optcb = cb;
   tv_free(tv);
   return OK;
-}
-
-static void didset_options_sctx(int opt_flags, int *buf)
-{
-  for (int i = 0;; i++) {
-    if (buf[i] == kOptInvalid) {
-      break;
-    }
-
-    set_option_sctx(buf[i], opt_flags, current_sctx);
-  }
 }
 
 /// Check if backspacing over something is allowed.
