+++ src/nvim/option_cb.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "nvim/option_defs.h"
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "option_cb.h.generated.h"
+#endif
